+++ client/src/components/pages/Store/Wallet/Transactions/index.js
@@ -1,0 +1,1 @@
+export { Transactions } from "./Transactions";
